Refactor Dictionary to use u32 for IDs and update related methods
- Changed ID types in Dictionary from u64 to u32 for memory efficiency.
- Updated encode and decode methods to reflect the new ID type.
- Adjusted tests to use the new encoding/decoding methods.
- Modified memory_tracker to track memory usage with detailed statistics.
- Added MemoryTracker struct for monitoring memory usage during runtime.
- Implemented methods for recording, retrieving, and resetting memory measurements.
- Enhanced segmented storage to utilize Rc for dictionary management.
- Updated utility functions to include a new StorageComponentSizes struct for memory breakdown.

+++ benches/benchmark.rs
@@ -16,10 +16,10 @@ fn setup_data(number_records: u64) -> std::io::Result<()> {
 
     for i in 0..number_records {
         let timestamp = i;
-        let subject = (i % 1000) as u64;
-        let predicate = (i % 500) as u64;
-        let object = (i % 2000) as u64;
-        let graph: u64 = 1;
+        let subject = (i % 1000) as u32;
+        let predicate = (i % 500) as u32;
+        let object = (i % 2000) as u32;
+        let graph: u32 = 1;
         writer.append_record(timestamp, subject, predicate, object, graph)?;
     }
